change package name to ProvIt.
I'm about to make some rather big changes (make every node a blank node).
This might break something or be stupid. Therefore this commit.
The problem with the current concept is, that it does not allow for
multiple executions of the same process. At least not in a comfortable way.